@@ -69,7 +69,7 @@
     "prettier": "^2.4.1",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "^4.2.3",
+    "typescript": "^4.3.5",
     "typescript-eslint": "0.0.1-alpha.0"
   }
 }
