@@ -1,9 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { OverviewComponent } from './overview/overview.component';
-import { SplashComponent } from './splash/splash.component';
 import { LandingComponent } from './landing/landing.component';
-import { environment } from '../environments/environment';
 
 const routes: Routes = [
   {
@@ -12,7 +10,7 @@ const routes: Routes = [
   },
   {
     path: '',
-    component: environment.useDataServer ? LandingComponent : SplashComponent
+    component: LandingComponent
   },
   { path: '**', redirectTo: '', pathMatch: 'full' }
 ];
