@@ -1,10 +1,14 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { colours, facetNames } from '../_data';
+import { filterList } from '../_helpers';
 import {
   ColourSeriesData,
   CompareData,
   CompareDataDescriptor,
   HeaderNameType,
+  IHashNumber,
+  SortBy,
+  SortInfo,
   TableRow
 } from '../_models';
 
@@ -21,6 +25,7 @@ export class SnapshotsComponent {
   @Input() set facetName(facetName: string) {
     this._facetName = facetName;
 
+    // unapply all other facets
     Object.keys(this.compareDataAllFacets).forEach((fName: string) => {
       if (fName !== facetName) {
         const cd = this.compareDataAllFacets[fName];
@@ -53,89 +58,100 @@ export class SnapshotsComponent {
     });
   }
 
-  /** getNextAvailableColourIndex
-  /* calculates the next available colour by looking at which have been used
-  /* within the current facet data
-  /* @param { string : facetName } - the active facet
-  /* @returns number
-  */
-  getNextAvailableColourIndex(facetName: string): number {
-    const cd = this.compareDataAllFacets[facetName];
+  preSortAndFilter(
+    facetName: string,
+    seriesKeys: Array<string>,
+    sortInfo: SortInfo,
+    filterTerm = ''
+  ): void {
+    seriesKeys.forEach((seriesKey: string) => {
+      const cd = this.compareDataAllFacets[facetName][seriesKey];
+      const data = cd.data;
 
-    const usedIndexes = Object.keys(cd)
-      .filter((key: string) => {
-        return cd[key]['applied'];
-      })
-      .map((key: string) => {
-        return cd[key]._colourIndex;
-      });
+      let operandA;
+      let operandB;
+      let sortedKeys: Array<string>;
 
-    const res = new Array(colours.length)
-      .fill(null)
-      .findIndex((_, x: number) => {
-        return !usedIndexes.includes(x);
-      });
-    return res > -1 ? res : 0;
+      const sortByCount = sortInfo.by === SortBy.count;
+
+      if (cd.orderOriginal && sortInfo.dir === 0) {
+        sortedKeys = filterList(filterTerm, cd.orderOriginal);
+      } else {
+        sortedKeys = filterList(filterTerm, Object.keys(data)).sort(
+          (a: string, b: string) => {
+            if (sortByCount) {
+              operandA = data[a];
+              operandB = data[b];
+            } else {
+              operandA = a;
+              operandB = b;
+            }
+            if (sortInfo.dir === 1) {
+              return operandA > operandB ? 1 : operandA === operandB ? 0 : -1;
+            } else if (sortInfo.dir === -1) {
+              return operandA < operandB ? 1 : operandA === operandB ? 0 : -1;
+            } else {
+              return 0;
+            }
+          }
+        );
+      }
+      cd.orderPreferred = sortedKeys;
+    });
   }
 
-  /** applySeries
-  /* sets "applied" on series with identified by seriesKeys
-  /* assign colour indexes
+  /** getSeriesDataForChart
   /* @param { string : facetName } - the active facet
   /* @param { Array<string> : seriesKeys } - the keys of the series to apply
   /* @param { boolean : percent } - percent value switch
   /* @returns Array<ColourSeriesData> converts to data (for chart)
   */
-  applySeries(
+  getSeriesDataForChart(
     facetName: string,
     seriesKeys: Array<string>,
     percent: boolean,
-    reuseColours = false
+    offset: number,
+    maxRows: number
   ): Array<ColourSeriesData> {
-    return seriesKeys.map((seriesKey: string) => {
+    return seriesKeys.map((seriesKey: string, keyIndex: number) => {
       const cd = this.compareDataAllFacets[facetName][seriesKey];
-
-      let colourIndex = cd._colourIndex;
-
-      if (!reuseColours) {
-        colourIndex = this.getNextAvailableColourIndex(facetName);
-        cd._colourIndex = colourIndex;
-      }
-
-      cd.applied = true;
-
-      const csd: ColourSeriesData = {
-        data: percent ? cd.dataPercent : cd.data,
-        colour: colours[colourIndex],
+      const data = percent ? cd.dataPercent : cd.data;
+      const cdKeys = cd.orderPreferred.slice(offset, offset + maxRows);
+      return {
+        data: cdKeys.reduce((map: IHashNumber, pref: string) => {
+          map[`${pref} `] = data[pref];
+          return map;
+        }, {}),
+        colour: colours[keyIndex],
         seriesName: seriesKey
       };
-      return csd;
     });
   }
 
-  /** getSeriesDataForTable
+  /** getSeriesDataForGrid
   /* converts to grouped / interplated data for table
   **/
-  getSeriesDataForTable(
+  getSeriesDataForGrid(
     facetName: string,
     seriesKeys: Array<string>
   ): Array<TableRow> {
-    const allKeysInAllSeries: { [groupName: string]: true } = {};
+    const allPreferred: Array<string> = [];
     const result: Array<TableRow> = [];
     const cds = this.compareDataAllFacets[facetName];
 
     // sort series keys (by pinIndex)
-
     this.getSortKeys(seriesKeys);
 
     // collect all possible category values and assign totals
 
-    seriesKeys.forEach((seriesKey: string) => {
+    seriesKeys.forEach((seriesKey: string, keyIndex: number) => {
       const cd = cds[seriesKey];
 
-      Object.keys(cd.data).forEach((key: string) => {
-        allKeysInAllSeries[key] = true;
+      cd.orderPreferred.forEach((key: string) => {
+        allPreferred.push(key);
       });
+
+      cd._colourIndex = keyIndex;
 
       result.push({
         name: 'Total', // name will not be shown in the grid
@@ -148,8 +164,7 @@ export class SnapshotsComponent {
     });
 
     // interpolate rows: check all series for category value's groupKey - add row if it has data
-
-    Object.keys(allKeysInAllSeries).forEach((groupKey: string) => {
+    allPreferred.forEach((groupKey: string) => {
       seriesKeys.forEach((seriesKey: string) => {
         const cd = cds[seriesKey];
         const count = cd.data[groupKey];
@@ -174,8 +189,13 @@ export class SnapshotsComponent {
   snap(facetName: string, key: string, cdd: CompareDataDescriptor): void {
     const cd = this.compareDataAllFacets[facetName];
     cd[key] = cdd;
+
     cdd.pinIndex = this.pinIndex;
     this.pinIndex++;
+
+    // record the original order
+    cdd.orderOriginal = Object.keys(cdd.data).slice(0);
+    cdd.orderPreferred = Object.keys(cdd.data).slice(0);
 
     this.filteredCDKeys(facetName, 'applied').forEach((key: string) => {
       cd[key].applied = false;
@@ -229,9 +249,15 @@ export class SnapshotsComponent {
     }
   }
 
+  apply(facetName: string, seriesKeys: Array<string>): void {
+    seriesKeys.forEach((seriesKey: string) => {
+      this.compareDataAllFacets[facetName][seriesKey].applied = true;
+    });
+  }
+
   unapply(seriesKey: string): void {
     const cd = this.compareDataAllFacets[this._facetName][seriesKey];
-    cd._colourIndex = 0;
+    cd._colourIndex = null;
     cd.applied = false;
   }
 }
